commons: move things into their own libraries
